pngs2cel: add line offsets to slab header
The slab header is 10 bytes long, and holds five uint16 values which
specify the frame offset at which to find the pixel data of row 0, 32,
64, 96 and 128 (in y-direction). This information is used by scrollrt
to avoid decoding pixel data that will not be rendered due to clipping.

NOTE: the implementation has not been tested (as I lack tools which
make use of the slab header).

Updates #2.